pyproject.toml
[project]
name = "langgraph-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-cli[inmem]>=0.4.12",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
